Move VM testing framework to subfolder
