+++ src/Tests/dotnet-MsiInstallation.Tests/Framework/RemoteDirectory.cs
@@ -8,7 +8,7 @@ using System.Text;
 using System.Threading.Tasks;
 using FluentAssertions.Execution;
 
-namespace Microsoft.DotNet.MsiInstallerTests
+namespace Microsoft.DotNet.MsiInstallerTests.Framework
 {
     abstract class RemoteDirectory
     {
