+++ src/Tests/dotnet-MsiInstallation.Tests/Framework/VMAction.cs
@@ -10,7 +10,7 @@ using System.Text;
 using System.Threading.Tasks;
 using Microsoft.DotNet.Cli.Utils;
 
-namespace Microsoft.DotNet.MsiInstallerTests
+namespace Microsoft.DotNet.MsiInstallerTests.Framework
 {
     abstract class VMAction
     {
